requirements.txt
scikit-learn==1.4.0
fastapi==0.111.0
uvicorn==0.30.1
imbalanced-learn==0.12.3
joblib==1.4.2
openai==1.73.0
python-dotenv==1.1.0
matplotlib==3.10.1
seaborn==0.13.2
python-multipart==0.0.20
numpy>=1.24.0
pandas>=2.0.0